requirements.txt
asgiref==3.3.1
backcall==0.2.0
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.3.1
decorator==4.4.2
defusedxml==0.7.0rc2
Django==3.1.5
django-extensions==3.1.0
django-phone-field==1.8.1
django-rest-registration==0.5.6
django-templated-mail==1.1.1
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
djoser==2.1.0
flake8==3.8.4
idna==2.10
ipython==7.19.0
ipython-genutils==0.2.0
itypes==1.2.0
jedi==0.18.0
Jinja2==2.11.2
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
parso==0.8.1
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.10
ptyprocess==0.7.0
pycodestyle==2.6.0
pycparser==2.20
pyflakes==2.2.0
Pygments==2.7.4
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sqlparse==0.4.1
traitlets==5.0.5
typing==3.7.4.3
uritemplate==3.0.1
urllib3==1.26.2
wcwidth==0.2.5
